requirements.txt
sqlalchemy
sanic
sanic-restful
pyjwt
pytest
pytest-sanic
pytest-ordering